Class renaming getInventorySuccessfulTest and Logic readme Update
